requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.7.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
mysql-connector-python==9.0.0
mysqlclient==2.2.4
openai==1.51.2
outcome==1.3.0.post0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
